duck-cnn-c/scripts/requirements.txt
torch>=2.0
torchvision>=0.15
numpy>=1.24
scikit-learn>=1.3
Pillow>=9.5
tqdm>=4.66
